requirements.txt
anyio==3.6.2
bson==0.5.10
click==8.1.3
dnspython==2.3.0
fastapi==0.95.0
h11==0.14.0
idna==3.4
joblib==1.2.0
movie-engine==0.1.2
numpy==1.24.2
pandas==2.0.0
pydantic==1.10.7
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.21.1
